Backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pillow
cssutils
jsbeautifier
beautifulsoup4
python-multipart
